+++ woocommerce/loop/add-to-cart.php
@@ -10,10 +10,10 @@
  * happen. When this occurs the version of the template file will be bumped and
  * the readme will list any important changes.
  *
- * @see           https://docs.woocommerce.com/document/template-structure/
- * @author        WooThemes
- * @package       WooCommerce/Templates
- * @version       3.0.0
+ * @see         https://docs.woocommerce.com/document/template-structure/
+ * @author      WooThemes
+ * @package     WooCommerce/Templates
+ * @version     3.3.0
  */
 
 if ( ! defined( 'ABSPATH' ) ) {
@@ -23,13 +23,14 @@ if ( ! defined( 'ABSPATH' ) ) {
 global $product;
 
 echo apply_filters(
-	'woocommerce_loop_add_to_cart_link',
-	sprintf( '<a rel="nofollow" href="%s" data-quantity="%s" data-product_id="%s" data-product_sku="%s" class="%s"><span class="fa fa-shopping-cart"></span> %s</a>',
+	'woocommerce_loop_add_to_cart_link', // WPCS: XSS ok.
+	sprintf(
+		'<a href="%s" data-quantity="%s" class="%s" %s><span class="fa fa-shopping-cart"></span> %s</a>',
 		esc_url( $product->add_to_cart_url() ),
-		esc_attr( isset( $quantity ) ? $quantity : 1 ),
-		esc_attr( $product->get_id() ),
-		esc_attr( $product->get_sku() ),
-		esc_attr( isset( $class ) ? $class : 'button' ),
+		esc_attr( isset( $args['quantity'] ) ? $args['quantity'] : 1 ),
+		esc_attr( isset( $args['class'] ) ? $args['class'] : 'button' ),
+		isset( $args['attributes'] ) ? wc_implode_html_attributes( $args['attributes'] ) : '',
 		esc_html( $product->add_to_cart_text() )
 	),
-$product );
+	$product, $args
+);