+++ woocommerce/single-product.php
@@ -34,11 +34,14 @@ get_header( 'shop' ); ?>
 				do_action( 'woocommerce_before_main_content' );
 				?>
 
-				<?php while ( have_posts() ) : the_post(); ?>
+				<?php
+				while ( have_posts() ) :
+					the_post();
+?>
 
 					<?php wc_get_template_part( 'content', 'single-product' ); ?>
 
-				<?php endwhile; // end of the loop. ?>
+				<?php endwhile; ?>
 
 				<?php
 				/**
@@ -59,6 +62,7 @@ get_header( 'shop' ); ?>
 		do_action( 'woocommerce_sidebar' );
 		?>
 	</div>
-<?php get_footer( 'shop' );
+<?php
+get_footer( 'shop' );
 
 /* Omit closing PHP tag at the end of PHP files to avoid "headers already sent" issues. */